requirements.txt
fastapi[standard]
dotenv
uvicorn
asyncpg
pydantic-settings
sqlmodel
alembic
passlib
bcrypt